bigint for area new shapefile new center for the initial map

+++ code/mainsettings/userwebsite/models.py
@@ -33,7 +33,7 @@ class Riversides(models.Model):
     """
 
     MYFLD = models.IntegerField()
-    AREA = models.IntegerField()
+    AREA = models.BigIntegerField()
 
     poly = models.MultiPolygonField(geography=False)
     #poly = models.PolygonField(geography=False)
